Implement height scaling when shrunk in Duke
also affects stereo separation

+++ source/games/blood/src/blood.cpp
@@ -51,7 +51,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include "tilesetbuilder.h"
 #include "nnexts.h"
 
-extern int playerHeight; //Used to define player height for VR
+extern float playerHeight; //Used to define player height for VR
 
 BEGIN_BLD_NS
