+++ source/games/duke/src/flags_r.cpp
@@ -30,7 +30,7 @@ Prepared for public release, 03/21/2003 - Charlie Wiederhold, 3D Realms
 #include "global.h"
 #include "names_r.h"
 
-extern int playerHeight;
+extern float playerHeight;
 
 BEGIN_DUKE_NS
